Separete functions into more semantic files and pipelines are working now
